// src/lib/fineract/generated/types.gen.ts
export type GetRolesResponse = {
    description?: string;
    id?: number;
    name?: string;
};